sift+kmeans/requirements.txt
streamlit
opencv-python-headless
numpy
scikit-learn
tensorflow
joblib
Pillow
